주차장실시간API(작업중)

+++ src/main/java/com/iot/parkingAPI/parkingAPIServiceImpl.java
@@ -54,8 +54,8 @@ public class parkingAPIServiceImpl implements parkingAPIService {
 				
 				String QUE_STATUS = (String)rowObject.get("QUE_STATUS");
 				String QUE_STATUS_NM = (String)rowObject.get("QUE_STATUS_NM");
-				int CAPACITY = Integer.parseInt((String)rowObject.get("CAPACITY"));
-				int CUR_PARKING = Integer.parseInt((String)rowObject.get("CUR_PARKING"));
+				double CAPACITY = Integer.parseInt((String)rowObject.get("CAPACITY"));
+				double CUR_PARKING = Integer.parseInt((String)rowObject.get("CUR_PARKING"));
 				String CUR_PARKING_TIME = (String)rowObject.get("CUR_PARKING_TIME");
 				String PARKING_CODE = (String)rowObject.get("PARKING_CODE");
